# server/swagger.json
{
  "openapi": "3.0.2",
  "info": {
...
    "version": "1.0.0",
    "title": "CashStory API (aka Darknight)",
    "termsOfService": "https://cashstory.com/en/legal",
    "contact": {
      "email": "bob@cashstory.com"
    }
  },
...
  "paths": {
...
    "/workspaces/{workspaceId}/section/{sectionId}/box": {
      "post": {
        "tags": [
          "workspace box"
        ],
        "summary": "Add new box in a section",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "workspace.addNewBoxToExistingSection",
        "parameters": [
          {
            "in": "path",
            "name": "workspaceId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sectionId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IBox"
              }
            }
          },
          "description": "Add box object",
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IBox"
                }
              }
            }
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/section/{sectionId}/box/{boxId}": {
      "put": {
        "tags": [
          "workspace box"
        ],
        "summary": "Update workspace box by id",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "workspace.updateExistingBox",
        "parameters": [
          {
            "in": "path",
            "name": "workspaceId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sectionId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "boxId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IBox"
              }
            }
          },
          "description": "Updated box object",
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IBox"
                }
              }
            }
          },
          "400": {
            "description": "error"
          }
        }
      },
      "delete": {
        "tags": [
          "workspace box"
        ],
        "summary": "Delete workspace box by id",
        "security": [
          {
            "BearerAuth": [
              "admin"
            ]
          }
        ],
        "description": "",
        "operationId": "workspace.deleteBoxFromSection",
        "parameters": [
          {
            "in": "path",
            "name": "workspaceId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sectionId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "boxId",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          },
          "400": {
            "description": "error"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "BasicAuth": [
        "read",
        "write"
      ]
    },
    {
      "BearerAuth": [
        "read",
        "write"
      ]
    }
  ],
  "servers": [
    {
      "url": "https://{environment}.cashstory.com/api/v1",
      "variables": {
        "environment": {
          "default": "darkknight",
          "enum": [
            "darkknight",
            "darkknight.dev"
          ]
        }
      }
    }
  ],
  "components": {
...
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "ILogin": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "IBox": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "urlBg"
        ],
        "properties": {
          "_id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "urlBg": {
            "type": "string"
          },
          "iframe": {
            "type": "boolean"
          },
          "hideElements": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "login": {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/ILogin"
              }
            }
          },
          "smartTable": {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/ISmartTable"
              }
            }
          },
          "cors": {
            "type": "string"
          },
          "class": {
            "type": "string"
          },
          "buttons": {
            "type": "object",
            "properties": {
              "notebookUrl": {
                "type": "string"
              },
              "expandCollapse": {
                "type": "boolean"
              },
              "openNewWindow": {
                "type": "boolean"
              },
              "backTo": {
                "type": "boolean"
              }
            }
          },
          "authMethod": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "autoExpand": {
            "type": "boolean"
          }
        },
        "xml": {
          "name": "IBox"
        }
      },
...
      "ISmartTable": {
        "type": "object",
        "required": [
          "database",
          "collection",
          "settings"
        ],
        "properties": {
          "database": {
            "type": "string"
          },
          "collection": {
            "type": "string"
          },
          "settings": {
            "type": "object"
          },
          "model": {
            "type": "string"
          },
          "rights": {
            "type": "object",
            "properties": {
              "items": {
                "$ref": "#/components/schemas/IRights"
              }
            }
          }
        },
        "xml": {
          "name": "ISmartTable"
        }
      },
...
      "IRights": {
        "type": "object",
        "properties": {
          "items": {
            "type": "string",
            "enum": [
              "GET",
              "PUT",
              "DELETE",
              "POST"
            ]
          }
        }
      },
...
    }
  }
}